Fixes and Unit tests
Wrote unit tests for the GildedRose.cs. While doing so found some errors to the "Aged Brie", it would be able to go over 50. To make it DRY I added a check to see if any item went over 50 and is not the infamous "Sulfuras" item(because both "Aged Brie" and "Backstage passes" can go over the 50 and shouldn't). Also discovered that "Conjured" item has to go twice as fast as the regular items so once the sell by date has passed it should go down in quality for -4 instead of regular -2.