Finalizing with some comments
Wrote some comments in code, so it's easier to understand what is happening. The task didn't ask to make it so the system could take "Backstage passes" of let's say different concerts. It could probably be possible to implement it so the switch only takes the beginning of the phrase "Backstage passes" and do the same logic the "Backstage passes" should have. The same would go to "Conjured" items, maybe even "Aged", cause aged could have the same logic. But then again this was not specified, but it would make sense.